app/requirements.txt
langchain==0.0.340
streamlit==1.28.2
azure-identity
python-dotenv==1.0.0
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
openai==1.3.7